Change type of getGeak to double

+++ src/main/java/ru/sokolova/jungleSimulator/model/Parrot.java
@@ -18,7 +18,7 @@ public class Parrot {
         return this.energy;
 
     }
-    public int getBeak(){
+    public double getBeak(){
         return this.energyProductionCoefficient;
     }
